+++ fugerit-springboot3-template/src/main/java/org/fugerit/java/template/fugerit_springboot3_template/FugeritSpringboot3TemplateApplication.java
@@ -1,13 +1,13 @@
-package org.fugerit.java.template.fugerit_springboot_template;
+package org.fugerit.java.template.fugerit_springboot3_template;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class FugeritSpringbootTemplateApplication {
+public class FugeritSpringboot3TemplateApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(FugeritSpringbootTemplateApplication.class, args);
+		SpringApplication.run(FugeritSpringboot3TemplateApplication.class, args);
 	}
 
 }